small fixes + addition of notifications.
Added a few checks to avoid the game crashing due to a stray null value.
Added a notification system based on the minecraft toasts (used for achievements, certain system messages, tutorial, recipes).

Fixes a bug with the location file not actually being loaded.

Cache no longer tries to update itself when the file doesn't exist (will create an "empty" cache file in its place).

Added in game notifications when the player enters / leaves an area defined in the location cache.