Reynart delve
Tells you which dungeon your delve bounty equates to.

Added "delveBountyMessage" to shards which is a non-necessary parameter that is used by the new feature to detect what's the player's current delve bounty.

Delve bounty recognition part 2.

Added an option to disable delve bounty recognition.

Switched delve bounty recognition to be based on localized text instead of only supporting english translatableText format: "unofficial-monumenta-mod.delvebounty.$shardName", it also requires "canBeDelveBounty" to be set to true for the shard.

Fixed an issue where changing the loaded shard wouldn't load the shard in locations.

Also it's 1.8.2 worthy imo.

THIS COMMIT IS TO RE-SYNCHRONIZE MY CURRENT CHANGES WITH THE UPSTREAM ONES