Ability Duration
Tells you which dungeon your delve bounty equates to.

Added "delveBountyMessage" to shards which is a non-necessary parameter that is used by the new feature to detect what's the player's current delve bounty.

Delve bounty recognition part 2.

Added an option to disable delve bounty recognition.

Switched delve bounty recognition to be based on localized text instead of only supporting english translatableText format: "unofficial-monumenta-mod.delvebounty.$shardName", it also requires "canBeDelveBounty" to be set to true for the shard.

Fixed an issue where changing the loaded shard wouldn't load the shard in locations.

Also it's 1.8.2 worthy imo.

Negative effects that are above 0 are now in red instead of in green

If an effect is stackable and it's strength is equal to 0, it will not show up in the overlay anymore.

Added an error if the parsing of a shard in /ummShard is unable to find anything

Also moved .effect to .effects because intelliJ thought .effect.Effect was inaccessible

Moved SlotLocking#DrawPolygon to Utils#drawFilledPolygon as well as its partial counterpart.

Added utility method Utils#drawHollowPolygon as well as its partial counterpart.

The mod now has the ability to receive durations through packets and visually convey them to the user.

Moved ChestCountOverlay to hud package.

Added Circle and Bar rendering (bar should currently not be used and is just a placeholder)