Fix an issue with shard checks
Since the change from String to TabShard it seems the equality check to check if the current shard was the same as the last shard was checking for reference equality instead of value equality.

Limited Teleport packet catching to 1 every second.

Moved shard debug stuff to a config value "shardDebug".